+++ components/engine/include/polygon.hpp
@@ -28,7 +28,7 @@ public:
   int texwidth;            // width of texture
   int texheight;           // height of texture
 
-  double r, g, b; // for flat coloring
+  float r, g, b; // for flat coloring
 
   RenderType rType; // Method of rendering this polygon
 
@@ -38,11 +38,11 @@ public:
 
   // THESE ARE FOR RENDERING QUICKLY
   Vertex ySorted[POLY_MAX_VERTICES]; // may not need this; just sort v[]
-  double increments[POLY_MAX_VERTICES][2]
-                   [NUM_VERTEX_DATA]; // increments along edges of left & right slopes
-  int edges[POLY_MAX_VERTICES];       // maintains connectivity after sorting
-  int sides[POLY_MAX_VERTICES][2];    // indices for starting vertices of each edge
-  int numInterps;                     // should use this for further optimization
+  float increments[POLY_MAX_VERTICES][2]
+                  [NUM_VERTEX_DATA]; // increments along edges of left & right slopes
+  int edges[POLY_MAX_VERTICES];      // maintains connectivity after sorting
+  int sides[POLY_MAX_VERTICES][2];   // indices for starting vertices of each edge
+  int numInterps;                    // should use this for further optimization
 
   Poly() {
     numVertices = 0;
@@ -97,15 +97,15 @@ public:
   }
   void SetRenderType(const RenderType rt) { rType = rt; }
   void SetNormal(const Vector3D n) { normal = n; }
-  void SetColor(const double _r, const double _g, const double _b) {
+  void SetColor(const float _r, const float _g, const float _b) {
     r = _r;
     g = _g;
     b = _b;
   }
-  void SetVertexColors(const double _r1, const double _g1, const double _b1, const double _r2,
-                       const double _g2, const double _b2, const double _r3, const double _g3,
-                       const double _b3, const double _r4 = 1, const double _g4 = 1,
-                       const double _b4 = 1) {
+  void SetVertexColors(const float _r1, const float _g1, const float _b1, const float _r2,
+                       const float _g2, const float _b2, const float _r3, const float _g3,
+                       const float _b3, const float _r4 = 1, const float _g4 = 1,
+                       const float _b4 = 1) {
     v[0].r = _r1;
     v[0].g = _g1;
     v[0].b = _b1;
@@ -123,7 +123,7 @@ public:
   // General Transformation Methods, only operate on x,y,z,w of vertices
   void Transform(const Matrix &_m);
   void Translate(const Vector3D &_v);
-  void Translate(const double _x, const double _y, const double _z);
+  void Translate(const float _x, const float _y, const float _z);
 
   // Pipeline Transformation Methods
   void TransformToCamera(const Matrix &_m);
@@ -145,12 +145,12 @@ public:
   void YSort(Vertex *temp);
   void XSort(Vertex *temp);
   void ZSort(Vertex *temp);
-  double MinX();
-  double MinY();
-  double MinZ();
-  double MaxX();
-  double MaxY();
-  double MaxZ();
+  float MinX();
+  float MinY();
+  float MinZ();
+  float MaxX();
+  float MaxY();
+  float MaxZ();
 
   // Operator Overloads
   Poly &operator=(const Poly &rhs);
